\name{hivc.seq.read.GenBank}
\alias{hivc.seq.read.GenBank}
\title{this file contains all R functions of the hivclust package}
\usage{
  hivc.seq.read.GenBank(access.nb, seq.names = access.nb,
    species.names = TRUE, gene.names = FALSE,
    as.character = FALSE, attributes = c("origin"))
}
\description{
  this file contains all R functions of the hivclust
  package
}

